--- frontend/apps/wps-web/src/features/fba/context/MapContext.tsx
@@ -1,4 +1,0 @@
-import React from "react";
-import { Map } from "ol";
-
-export const MapContext = React.createContext<Map | null>(null);
