--- frontend/apps/wps-web/src/features/fba/fireZoneUnitExtents.ts
@@ -1,47 +1,0 @@
-// Generated with postGIS query:
-// SELECT source_identifier, ST_Extent(ST_Transform(geom, 3857)) AS bbox
-// FROM advisory_shapes
-// GROUP BY source_identifier
-
-export const fireZoneExtentsMap = new Map([
-  ['9', [-1.2990446453409778e7, 6274667.515325945, -1.2696450026696604e7, 6548411.326476511]],
-  ['8', [-1.4239393527168721e7, 6354692.628401642, -1.3885628823239345e7, 6614726.5444521075]],
-  ['7', [-1.3804737218859635e7, 6344805.590225804, -1.3585455009676876e7, 6593252.999027362]],
-  ['6', [-1.3507294147147883e7, 6274897.967954165, -1.334774839441576e7, 6537705.402908283]],
-  ['5', [-1.3409245893438928e7, 6274766.742380009, -1.3220235499035008e7, 6481949.463285208]],
-  ['40', [-1.5480234287311064e7, 7577642.392997014, -1.4112365951461505e7, 8400167.169926997]],
-  ['4', [-1.3751776569011472e7, 6274402.126349792, -1.3424911887489706e7, 6472043.652311066]],
-  ['39', [-1.4351962980972972e7, 7853828.243202188, -1.3358477755491769e7, 8400802.80211162]],
-  ['38', [-1.3827793521976218e7, 7576304.957512168, -1.3358479304175606e7, 8088009.006763926]],
-  ['37', [-1.4219800072231779e7, 7336065.278003179, -1.360854082549229e7, 8109561.9143451]],
-  ['36', [-1.4336781245129095e7, 7204442.79685761, -1.371169548188525e7, 7820500.999762873]],
-  ['35', [-1.450768537050546e7, 6924210.901583922, -1.416150856031427e7, 7798283.80675702]],
-  ['34', [-1.3693165305696633e7, 7189897.212373013, -1.3358474603134027e7, 7625578.3247031225]],
-  ['33', [-1.4342555408019697e7, 7325141.207085127, -1.4118247090158451e7, 7608285.9692941625]],
-  ['32', [-1.4264130993376458e7, 7251562.216603748, -1.4069094425244572e7, 7559646.143026362]],
-  ['31', [-1.465995968053417e7, 6791108.103214191, -1.4265017487204237e7, 7547089.7768202685]],
-  ['30', [-1.425624634417718e7, 7053861.425825395, -1.3984738982186943e7, 7463618.604611745]],
-  ['3', [-1.3280559041253906e7, 6274868.297736709, -1.3139838132862406e7, 6431584.604443101]],
-  ['29', [-1.3814924725325251e7, 7005134.697469937, -1.3358483054930253e7, 7402698.959496517]],
-  ['28', [-1.4178582016564626e7, 6977090.497498701, -1.3900327488148177e7, 7357120.120208108]],
-  ['27', [-1.4830653531238658e7, 6661969.272163259, -1.4438514540166311e7, 7303487.644342587]],
-  ['26', [-1.4001085155298334e7, 6967533.04054967, -1.373221467010641e7, 7256396.031430607]],
-  ['25', [-1.3473946106509669e7, 6800425.7877606945, -1.3157325389145028e7, 7133536.671827988]],
-  ['24', [-1.3980018237322483e7, 6868194.473228056, -1.3417375254083173e7, 7071337.897894138]],
-  ['23', [-1.4454616078859169e7, 6648018.459231251, -1.3979012866164453e7, 7039647.156581514]],
-  ['22', [-1.3594776668727398e7, 6798564.728984073, -1.3373863782521151e7, 6981899.065810441]],
-  ['21', [-1.4001084621837731e7, 6599377.316962451, -1.365754947442412e7, 6975295.812021961]],
-  ['20', [-1.3752430898725286e7, 6633727.7494424265, -1.353171475899193e7, 6930819.026353851]],
-  ['2', [-1.4102327737002596e7, 6190322.438244796, -1.3739154606700154e7, 6398124.051320654]],
-  ['19', [-1.3246186938489951e7, 6521438.3495329935, -1.289470341755373e7, 6890528.452262035]],
-  ['18', [-1.3613238848690292e7, 6613230.764917851, -1.3383569376688499e7, 6834856.728304669]],
-  ['17', [-1.4439282220115436e7, 6427871.498074888, -1.39374621994989e7, 6789082.8767122775]],
-  ['16', [-1.3551993942676567e7, 6500313.4256845005, -1.3230031995602995e7, 6960226.369767584]],
-  ['15', [-1.30612924356551e7, 6406284.152275828, -1.280735574931457e7, 6672210.129222219]],
-  ['14', [-1.318278203880094e7, 6274659.0787313515, -1.3025107275702618e7, 6641088.49173979]],
-  ['13', [-1.3086349552941706e7, 6274700.946005805, -1.288249024528788e7, 6634238.803499517]],
-  ['12', [-1.3773818093438104e7, 6450265.879861334, -1.3501267252202304e7, 6671039.04498325]],
-  ['11', [-1.395267371568761e7, 6313474.478500485, -1.372014151941535e7, 6679343.359690739]],
-  ['10', [-1.3370320686968567e7, 6415289.471154898, -1.3151381999755923e7, 6745929.533973716]],
-  ['1', [-1.3885884887820601e7, 6144341.120806807, -1.3693291829990938e7, 6304097.619536648]]
-])
