+++ frontend/apps/wps-web/src/features/sfmsInsights/components/map/sfmsFeatureStylers.ts
@@ -1,4 +1,4 @@
-import { getColorByFuelTypeCode } from '@/features/fba/components/viz/color'
+import { getColorByFuelTypeCode } from '@wps/asa/color'
 import { RASTER_COLOR_BREAKS, RasterType, ColorBreak, FUEL_TYPE_COLORS } from './rasterConfig'
 import * as ol from 'ol'
 import Geometry from 'ol/geom/Geometry'
