+++ frontend/packages/asa/src/fireZoneUnitExtents.ts
